requirements/cdc-enhanced.txt
# Enhanced requirements for 500+ table CDC with Spark 4
pyspark>=3.5.0
delta-spark>=3.0.0
mysql-connector-python>=8.0.0
pymysql>=1.0.0
pymysqlreplication>=0.30
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
prometheus-client>=0.16.0
structlog>=23.1.0
tenacity>=8.2.0
cachetools>=5.3.0
aiofiles>=23.1.0
aiomysql>=0.2.0
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0